Fixed issue with mass. Added vertet integration. Fixed leap frog integration.

+++ ParticlePhysics/Entities/RigidBody.cpp
@@ -61,7 +61,7 @@ void RigidBody::initCube(const real dimensions[], real particleRadius, const rea
 
   int index = 0;
 
-  const real perParticleInvMass = real(mass) / real(particleCount);
+  const real perParticleInvMass = real(particleCount) / real(mass);
 
   setInvMassIsShared((*entitySharedData.host())[0], true);
   (*entitySharedData.host())[0].sharedInvMass = perParticleInvMass;
